// frontend/src/api/api.ts
export type RegisterPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.RegisterPostSchema>
  }
  response: unknown
}
export type LoginPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.LoginPostSchema>
  }
  response: unknown
}
export type OauthGetProviderRedirectContract = {
  request: {
    params: { provider: string }
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.OauthGetProviderRedirectSchema>
  }
  response: unknown
}
export type OauthGetProviderCallbackContract = {
  request: {
    params: { provider: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type OauthPostProviderCallbackContract = {
  request: {
    params: { provider: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type SocialPostLoginContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.SocialPostLoginSchema>
  }
  response: unknown
}
export type ForgotPasswordPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.ForgotPasswordPostSchema>
  }
  response: unknown
}
export type ResetPasswordPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.ResetPasswordPostSchema>
  }
  response: unknown
}
export type CategoriesGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type ProdukGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: Types.ProdukItem[]
}
export type ProdukGetIdContract = {
  request: {
    params: { id: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: Types.ProdukItem
}
export type ProdukGetIdReviewsContract = {
  request: {
    params: { id: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type ProdukPostIdReviewsContract = {
  request: {
    params: { id: string }
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.ProdukPostIdReviewsSchema>
  }
  response: unknown
}
export type PaymentPostWebhookContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type PaymentPostOrderIdContract = {
  request: {
    params: { orderId: string }
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.PaymentPostOrderIdSchema>
  }
  response: unknown
}
export type ProfileGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type ProfilePutContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.ProfilePutSchema>
  }
  response: unknown
}
export type ProfilePatchContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.ProfilePatchSchema>
  }
  response: unknown
}
export type OrdersGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type OrdersGetIdContract = {
  request: {
    params: { id: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type OrdersGetIdInvoiceContract = {
  request: {
    params: { id: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type CartPostItemsContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.CartPostItemsSchema>
  }
  response: unknown
}
export type CartPatchItemsProdukItemIdContract = {
  request: {
    params: { produkItemId: string }
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.CartPatchItemsProdukItemIdSchema>
  }
  response: unknown
}
export type CartDeleteItemsProdukItemIdContract = {
  request: {
    params: { produkItemId: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type CartDeleteContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type CartPostPromoContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.CartPostPromoSchema>
  }
  response: unknown
}
export type CartDeletePromoContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type CheckoutPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.CheckoutPostSchema>
  }
  response: unknown
}
export type BuyNowPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.BuyNowPostSchema>
  }
  response: unknown
}
export type KeranjangGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type WishlistGetContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: Types.ProdukItem[]
}
export type WishlistPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.WishlistPostSchema>
  }
  response: unknown
}
export type WishlistDeleteProdukItemIdContract = {
  request: {
    params: { produkItemId: string }
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type LogoutPostContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body?: unknown
  }
  response: unknown
}
export type AdminPostProdukContract = {
  request: {
    params?: unknown
    query?: Record<string, unknown>
    body: z.infer<typeof Schemas.AdminPostProdukSchema>
  }
  response: unknown
}

export const api = defineApi({
  register: {
    post: endpoint<RegisterPostContract['response'], RegisterPostContract['request']['params'], RegisterPostContract['request']['body']>({
      method: 'POST',
      path: '/register',
      schema: {
        body: Schemas.RegisterPostSchema
      },
    }),
  },
// ...
})